+++ src/components/journal/folder.tsx
